move non-exported files into a lib directory
