+++ src/jsx.js
@@ -1,6 +1,6 @@
-import './polyfills';
-import renderToString from './index';
-import { indent, encodeEntities } from './util';
+import './lib/polyfills';
+import renderToString from './pretty';
+import { indent, encodeEntities } from './lib/util';
 import prettyFormat from 'pretty-format';
 
 // we have to patch in Array support, Possible issue in npm.im/pretty-format
@@ -67,10 +67,36 @@ let defaultOpts = {
 	pretty: '  '
 };
 
-function renderToJsxString(vnode, context, opts, inner) {
-	opts = Object.assign({}, defaultOpts, opts || {});
-	return renderToString(vnode, context, opts, inner);
+/**
+ * Render Preact JSX + Components to a pretty-printed HTML-like string.
+ * @param {VNode} vnode	JSX Element / VNode to render
+ * @param {Object} [context={}] Initial root context object
+ * @param {Object} [options={}] Rendering options
+ * @param {Boolean} [options.jsx=true] Generate JSX/XML output instead of HTML
+ * @param {Boolean} [options.xml=false] Use self-closing tags for elements without children
+ * @param {Boolean} [options.shallow=false] Serialize nested Components (`<Foo a="b" />`) instead of rendering
+ * @param {Boolean} [options.pretty=false] Add whitespace for readability
+ * @param {RegExp|undefined} [options.voidElements] RegeEx to define which element types are self-closing
+ * @returns {String} a pretty-printed HTML-like string
+ */
+export default function renderToStringPretty(vnode, context, options) {
+	const opts = Object.assign({}, defaultOpts, options || {});
+	if (!opts.jsx) opts.attributeHook = null;
+	return renderToString(vnode, context, opts);
 }
+export { renderToStringPretty as render };
 
-export default renderToJsxString;
-export { renderToJsxString as render };
+const SHALLOW = { shallow: true };
+
+/** Only render elements, leaving Components inline as `<ComponentName ... />`.
+ *	This method is just a convenience alias for `render(vnode, context, { shallow:true })`
+ *	@name shallow
+ *	@function
+ *	@param {VNode} vnode	JSX VNode to render.
+ *	@param {Object} [context={}]	Optionally pass an initial context object through the render path.
+ *	@param {Parameters<typeof renderToStringPretty>[2]} [options]	Optionally pass an initial context object through the render path.
+ */
+export function shallowRender(vnode, context, options) {
+	const opts = Object.assign({}, SHALLOW, options || {});
+	return renderToStringPretty(vnode, context, opts);
+}
