+++ src/lib/polyfills.js
